Availability annotations don’t work for tests
Make sense since the runtime is used to find the tests, and that is not Swift-aware.

Can’t see any other way to do this.

Fixes CI for iOS 8 tests since seemingly they didn’t back port the stubs for this function that far and it then fails. Frankly I don’t get it. Since the iOS 9 tests passed.